refactor: slim training route records
